+++ gradle.properties
@@ -1,1 +1,1 @@
-version=0.0.1-SNAPSHOT
+version=0.0.12-SNAPSHOT
