+++ jmix-wizard/src/main/java/de/diedavids/jmix/wizard/impl/WizardButtonsPanel.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 
-public class WizardButtonsPanel {
+class WizardButtonsPanel {
 
     private final UiComponents uiComponents;
     private final Messages messages;
