+++ jmix-wizard/src/test/java/de/diedavids/jmix/wizard/ProgrammaticNavigationTest.java
@@ -1,47 +1,22 @@
-package de.diedavids.jmix.wizard.screen;
+package de.diedavids.jmix.wizard;
 
-import de.diedavids.jmix.wizard.Wizard;
-import de.diedavids.jmix.wizard.WizardConfiguration;
-import de.diedavids.jmix.wizard.WizardTestConfiguration;
-import de.diedavids.jmix.wizard.screen.sample.four_steps.FourStepsWizard;
-import de.diedavids.jmix.wizard.screen.sample.simple.SimpleWizard;
-import io.jmix.core.CoreConfiguration;
-import io.jmix.data.DataConfiguration;
-import io.jmix.eclipselink.EclipselinkConfiguration;
+import de.diedavids.jmix.wizard.screen.test_support.FourStepsWizard;
 import io.jmix.ui.Screens;
-import io.jmix.ui.UiConfiguration;
-import io.jmix.ui.component.Button;
-import io.jmix.ui.testassist.junit.UiTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.test.context.ContextConfiguration;
 
 import java.util.EventObject;
 
 import static de.diedavids.jmix.wizard.Wizard.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@UiTest(
-        // authenticatedUser = "admin", --> throws 'User admin not found'
-        mainScreenId = "TestWizardMainScreen",
-        screenBasePackages = "de.diedavids.jmix.wizard.screen.sample"
-)
-@ContextConfiguration(classes = {
-        CoreConfiguration.class,
-        DataConfiguration.class,
-        EclipselinkConfiguration.class,
-        UiConfiguration.class,
-
-        WizardConfiguration.class,
-        WizardTestConfiguration.class
-})
-class ProgrammaticNavigationTest {
+class ProgrammaticNavigationTest extends WizardUiTest {
 
     private Wizard wizard;
     private FourStepsWizard screen;
 
     @BeforeEach
-    void openSimpleWizard(Screens screens) {
+    void openScreen(Screens screens) {
 
         screen = screens.create(FourStepsWizard.class);
