+++ jmix-wizard/src/test/java/de/diedavids/jmix/wizard/WizardUiTest.java
@@ -1,0 +1,28 @@
+package de.diedavids.jmix.wizard;
+
+import io.jmix.core.CoreConfiguration;
+import io.jmix.data.DataConfiguration;
+import io.jmix.eclipselink.EclipselinkConfiguration;
+import io.jmix.ui.UiConfiguration;
+import io.jmix.ui.testassist.junit.UiTest;
+import org.springframework.test.context.ContextConfiguration;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@UiTest(
+        // authenticatedUser = "admin", --> throws 'User admin not found'
+        mainScreenId = "TestWizardMainScreen",
+        screenBasePackages = "de.diedavids.jmix.wizard.screen.sample"
+)
+@ContextConfiguration(classes = {
+        CoreConfiguration.class,
+        DataConfiguration.class,
+        EclipselinkConfiguration.class,
+        UiConfiguration.class,
+
+        WizardConfiguration.class,
+        WizardTestConfiguration.class
+})
+class WizardUiTest {
+
+}
