+++ jmix-wizard/src/test/java/de/diedavids/jmix/wizard/screen/test_support/FourStepsWizard.java
@@ -1,0 +1,12 @@
+package de.diedavids.jmix.wizard.screen.test_support;
+
+
+import io.jmix.ui.screen.UiController;
+import io.jmix.ui.screen.UiDescriptor;
+
+
+@UiController("ddcw_FourStepsWizard")
+@UiDescriptor("four-steps-wizard.xml")
+public class FourStepsWizard extends AbstractWizardScreen {
+
+}
