+++ jmix-wizard/src/test/java/de/diedavids/jmix/wizard/screen/test_support/LazyWizard.java
@@ -1,0 +1,11 @@
+package de.diedavids.jmix.wizard.screen.test_support;
+
+
+import io.jmix.ui.screen.UiController;
+import io.jmix.ui.screen.UiDescriptor;
+
+@UiController("ddcw_LazyWizard")
+@UiDescriptor("lazy-wizard.xml")
+public class LazyWizard extends AbstractWizardScreen {
+
+}
