+++ jmix-wizard/src/test/java/de/diedavids/jmix/wizard/screen/test_support/SimpleWizard.java
@@ -1,0 +1,10 @@
+package de.diedavids.jmix.wizard.screen.test_support;
+
+import io.jmix.ui.screen.UiController;
+import io.jmix.ui.screen.UiDescriptor;
+
+@UiController("ddcw_SimpleWizard")
+@UiDescriptor("simple-wizard.xml")
+public class SimpleWizard extends AbstractWizardScreen {
+
+}
