+++ jmix-wizard/src/test/resources/de/diedavids/jmix/wizard/screen/test_support/test-wizard-main-screen.xml
